Rename `closingTag` into `voidTag` like in the official w3c spec. Provide a `createHtmlTagObject` helper for plugin authors to create new tags.
