+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change History
 ==============
 
+v3.x
+---
+* Rename `closingTag` into `voidTag` like in the official w3c spec.
+* Provide a `createHtmlTagObject` helper for plugin authors to create new tags.
+
 v2.26.0
 ---
 * Allow plugins to add attributes without values to the `<script>` and `<link>` tags
